+++ src/tests/test_main_menu.py
@@ -3,6 +3,7 @@ import pygame
 from unittest.mock import Mock, patch
 from src.classes.menu_state_enum import MenuStateEnum
 from src.helpers.main_menu import MainMenu
+from src.helpers.high_score_helper import HighScoreHelper
 
 @pytest.fixture
 def setup_pygame():
@@ -16,17 +17,18 @@ def screen():
 
 @pytest.fixture
 def main_menu(screen) -> MainMenu:
-    return MainMenu(screen)
+    high_score_helper = HighScoreHelper("")
+    return MainMenu(screen, high_score_helper)
 
 def test_initial_state(main_menu: MainMenu):
     assert main_menu.menu_state == MenuStateEnum.MAIN_MENU
     assert main_menu.selected_option == 0
 
-def test_handle_events_up_key_2(main_menu: MainMenu):
+def test_handle_events_up_key(main_menu: MainMenu):
     event = pygame.event.Event(pygame.KEYDOWN, {'key': pygame.K_UP})
     pygame.event.post(event)
     main_menu.handle_events()
-    assert main_menu.selected_option == 3
+    assert main_menu.selected_option == 4
 
 
 def test_handle_events_down_key(main_menu: MainMenu):
@@ -56,12 +58,18 @@ def test_handle_events_enter_key_bindings(main_menu: MainMenu):
     main_menu.handle_events()
     assert main_menu.menu_state == MenuStateEnum.KEY_BINDINGS
 
-def test_handle_events_enter_exit(main_menu: MainMenu):
+def test_handle_events_enter_view_high_score(main_menu: MainMenu):
     main_menu.selected_option = 3
+    event = pygame.event.Event(pygame.KEYDOWN, {'key': pygame.K_RETURN})
+    pygame.event.post(event)
+    main_menu.handle_events()
+    assert main_menu.menu_state == MenuStateEnum.VIEW_HIGH_SCORES
+
+def test_handle_events_enter_exit(main_menu: MainMenu):
+    main_menu.selected_option = 4
     event = pygame.event.Event(pygame.KEYDOWN, {'key': pygame.K_RETURN})
     with patch('pygame.quit') as mock_quit:
         with pytest.raises(SystemExit):
             pygame.event.post(event)
             main_menu.handle_events()
-            print(main_menu.menu_state)
             mock_quit.assert_called_once()